Move proto conversions into package
