+++ pkg/proto/conv/placement_state.go
@@ -1,4 +1,4 @@
-package ranje
+package conv
 
 import (
 	"fmt"
@@ -7,25 +7,6 @@ import (
 	"github.com/adammck/ranger/pkg/api"
 	pb "github.com/adammck/ranger/pkg/proto/gen"
 )
-
-func PlacementStateToProto(s api.PlacementState) pb.PlacementState {
-	switch s {
-	case api.PsUnknown:
-		return pb.PlacementState_PS_UNKNOWN
-	case api.PsPending:
-		return pb.PlacementState_PS_PENDING
-	case api.PsInactive:
-		return pb.PlacementState_PS_INACTIVE
-	case api.PsActive:
-		return pb.PlacementState_PS_ACTIVE
-	case api.PsGiveUp:
-		return pb.PlacementState_PS_GIVE_UP
-	case api.PsDropped:
-		return pb.PlacementState_PS_DROPPED
-	}
-
-	panic(fmt.Sprintf("unknown PlacementState: %#v", s))
-}
 
 func PlacementStateFromProto(s *pb.PlacementState) api.PlacementState {
 	switch *s {
@@ -45,4 +26,23 @@ func PlacementStateFromProto(s *pb.PlacementState) api.PlacementState {
 
 	log.Printf("warn: unknown PlacementState from proto: %v", *s)
 	return api.PsUnknown
+}
+
+func PlacementStateToProto(s api.PlacementState) pb.PlacementState {
+	switch s {
+	case api.PsUnknown:
+		return pb.PlacementState_PS_UNKNOWN
+	case api.PsPending:
+		return pb.PlacementState_PS_PENDING
+	case api.PsInactive:
+		return pb.PlacementState_PS_INACTIVE
+	case api.PsActive:
+		return pb.PlacementState_PS_ACTIVE
+	case api.PsGiveUp:
+		return pb.PlacementState_PS_GIVE_UP
+	case api.PsDropped:
+		return pb.PlacementState_PS_DROPPED
+	}
+
+	panic(fmt.Sprintf("unknown PlacementState: %#v", s))
 }
